Move various combat functions into combat modules
